requirements.txt
python-decouple
djangorestframework
pillow
django
djangorestframework_simplejwt
psycopg2-binary
dj-config-url
